pong socket improved
